feat(rbenv): gets ruby version from service
The idea is to have a endpoint to get the latests ruby version that we
are using at platanus. The endpoint is http://ruby.platan.us/latest

Potassium get the version the the service and generates the
corresponding `.ruby-version`

If the service is down, if fallback to whatever ruby version potassium
it's being run. This is done via the RBENV_VERSION environmental
variable.