feat(): Use .env files instead of .rbenv-vars
Why?

Heroku doesn't use rbenv to load diferent versions of ruby nor the
rbenv-vars plugin to load environmental variables.

Configuration in Heroku is done with real environmental varibles defined
using the GUI or CLI (`heroku config`)

In development the `.env` file is loaded in to different ways. With the
`dotenv` gem or if using `forego` or `heroku local` those tools makes
sure to load the `.env` into the environment